+++ src/hapi/pipelines/database/metadata.py
@@ -34,8 +34,8 @@ class Metadata(BaseUploader):
                 hdx_id=dataset_id,
                 hdx_stub=dataset["hdx_stub"],
                 title=dataset["title"],
-                provider_code=dataset["provider_code"],
-                provider_name=dataset["provider_name"],
+                hdx_provider_stub=dataset["hdx_provider_stub"],
+                hdx_provider_name=dataset["hdx_provider_name"],
             )
             self._session.add(dataset_row)
             self._session.commit()
@@ -53,11 +53,10 @@ class Metadata(BaseUploader):
                     if sheet["is_hxlated"]:
                         is_hxlated = True
                         break
-                # TODO: Change resource "filename" to "name" in hdx-python-scraper (HAPI-246)
                 resource_row = DBResource(
                     hdx_id=resource_id,
                     dataset_ref=dataset_row.id,
-                    name=resource["filename"],
+                    name=resource["name"],
                     format=resource["format"],
                     update_date=resource["update_date"],
                     is_hxl=is_hxlated,