+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.6.0] - 2023-10-26
+
+### Added
+
+- Minor unit tests
+- Food security and related tables for Burkina Faso, Chad, and Mali
+
+## [0.5.5] - 2023-10-19
+
+### Changed
+
+- Resource filename changed to name
+
+## [0.5.4] - 2023-10-19
+
+### Changed
+
+- HDX provider code and name change
+
 ## [0.5.3] - 2023-10-19
 
 ### Fixed